backend/requirements.txt
Django==3.2.3
django_filter==23.2
python-dotenv==1.0.0
